+++ apps/benchmarks/kernel/lavamd.c
@@ -1,0 +1,1 @@
+../../lavamd/kernel/lavamd.c